Fix: resolve g.2 test-quality findings and sync story artifacts
